refactor: move ReplicationService interface out of root package
